debugged output texture, reorganized c# script
currently 128x128 sim with 370 particles takes 1ms. but it seems like at least 0.5s is waiting on gpu.
the resolution and particle number both seem to significantly effect the sim. An acceleration data structure will very likely be needed... or to take more work off the gpu. perhap's the set data as well?